% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_project.R
\name{add_analysis}
\alias{add_analysis}
\title{Add analysis file from template}
\usage{
add_analysis(x = ".", name = NULL)
}
\arguments{
\item{x}{Character. Location at which to add a the file.}

\item{name}{Character. Name of the to-be-created file. Defaults to \code{analysis.Rmd}}
}
\description{
Adds an new R Markdown file at the specified location
}
