% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_project.R
\name{add_paper}
\alias{add_paper}
\title{Add a new paper}
\usage{
add_paper(x = ".", shorttitle)
}
\arguments{
\item{x}{Character. Path to project repository.}

\item{shorttitle}{Character. Short title of the paper used as directory name.}
}
\description{
Set up directory structure for a new paper and adds a \pkg{papaja} template to
an existing project repository.
}
