% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_project.R
\name{add_readme}
\alias{add_readme}
\title{Add README from template}
\usage{
add_readme(x = ".", name = NULL)
}
\arguments{
\item{x}{Character. Location at which to add a the file.}

\item{name}{Character. Name of the to-be-created file. Defaults to \code{README.Rmd}}
}
\value{

}
\description{
Adds a README.Rmd file at the specified location
}
