% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{author_from_desc}
\alias{author_from_desc}
\title{Get first author name}
\usage{
author_from_desc(...)
}
\arguments{
\item{...}{Passed to \code{desc::desc_get}}
}
\description{
Extracts first author name from DESCRIPTION
}
