% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{do_not_serialize}
\alias{do_not_serialize}
\title{Exit serialization from plumber}
\usage{
do_not_serialize(val, req, res, errorHandler)
}
\arguments{
\item{val}{val to serialize}

\item{req}{plumber req}

\item{res}{plumber res}

\item{errorHandler}{custom error handler}
}
\description{
Exit serialization from plumber
}
