% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texTools-preamble-biblatex-dot.R
\name{.PreambleBiblatex}
\alias{.PreambleBiblatex}
\title{Create Biblatex Preamble}
\usage{
.PreambleBiblatex(
  path = "bib",
  style = "apa",
  sortcites = TRUE,
  sorting = "nyt",
  map = TRUE
)
}
\arguments{
\item{path}{Character string.
Path to \verb{*.bib} files.}

\item{style}{Character string.
\code{"apa"} or \code{"reading"}.}

\item{sortcites}{Logical.
Sort citations.}

\item{sorting}{Character string.
\code{"nty"} sort by name, title, year;
\code{"nyt"} sort by name, year, title;
\code{"nyvt"} sort by name, year, volume, title;
\code{"anyt "} sort by alphabetic label, name, year, title;
\code{"anyvt"} sort by alphabetic label, name, year, volume, title;
\code{"ydnt"} sort by year (descending), name, title; and
\code{"none"} entries are processed in citation order.}

\item{map}{Logical.
Null mapping for \code{addendum}, \code{note}, and \code{annotation}.}
}
\value{
Character string.
}
\description{
Create Biblatex Preamble
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Bibliography Functions}
\keyword{format}
\keyword{internal}
\keyword{texTools}
