% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{examplefunction}
\alias{examplefunction}
\title{How to document functions.}
\usage{
examplefunction(text)
}
\arguments{
\item{text}{Message to print.}
}
\description{
This is an example showing how to document functions to be included in the package.
}
\examples{
examplefunction('Hello world!')
}
\author{
Francisco Rodriguez-Sanchez
}
\seealso{
\url{http://r-pkgs.had.co.nz/man.html}
}

