% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template-foo.R
\name{Foo}
\alias{Foo}
\title{Foo}
\usage{
Foo(x, y)
}
\arguments{
\item{x}{Numeric.}

\item{y}{Numeric.}
}
\description{
Sum function.
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Template Functions}
\keyword{template}
