% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_get.R
\name{get_this_from_that}
\alias{get_this_from_that}
\title{Get this from that}
\usage{
get_this_from_that(
  this_1L_chr = "first",
  that_ls = list(first = 5, second = 6)
)
}
\arguments{
\item{this_1L_chr}{This (a character vector of length one), Default: 'first'}

\item{that_ls}{That (a list), Default: list(first = 5, second = 6)}
}
\value{
This (an output object of multiple potential types)
}
\description{
get_this_from_that() is a Get function that extracts data from an object. Specifically, this function implements an algorithm to get this from that. The function returns This (an output object of multiple potential types).
}
