% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{init_drake}
\alias{init_drake}
\title{Initialize \pkg{drake} workflow}
\usage{
init_drake(x = ".", git, pkg_structure)
}
\arguments{
\item{x}{Character. Location at which to initialize drake workflow.}

\item{git}{Logical. Whether to initialize git.}

\item{pkg_structure}{Logical. Whether to add R-package infrastructure.}
}
\description{
Sets up build files and creates a dedicated \pkg{drake} cache.
}
