% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_project.R
\name{init_project}
\alias{init_project}
\title{Set up new project repository}
\usage{
init_project(
  x,
  path = ".",
  pkg_structure = TRUE,
  fields = NULL,
  git = TRUE,
  drake = FALSE,
  targets = FALSE,
  docker = TRUE
)
}
\arguments{
\item{x}{Character. Name of project repository.}

\item{path}{A path. If it exists, it is used. If it does not exist, it is
created, provided that the parent path exists.}

\item{pkg_structure}{Logical. Whether to add R-package infrastructure.}

\item{fields}{A named list of fields to add to \code{DESCRIPTION}, potentially
overriding default values. See \code{\link[usethis:use_description]{use_description()}} for how you can set
personalized defaults using package options}

\item{git}{Logical. Whether to initialize git.}

\item{drake}{Logical. Whether to set up \pkg{drake} infrastructure.}

\item{targets}{Logical. Whether to set up \pkg{targets} infrastructure.}

\item{docker}{Logical. Whether to configure a default Docker container. For
details see the
\url[papaja manual]{http://frederikaust.com/papaja_man/tips-and-tricks.html#docker}}
}
\description{
Creates a folder structure and sets up basic functionality.
}
\details{
When using \pkg{drake} keep the following in mind: First, customize
  \code{_drake_config.R} (e.g., remove or complete \code{expose_imports()}) and choose
  a parallelization approach. Parallelization requires the \pkg{future} package. Local
  parallelization should otherwise work out of the box. For parallelization on the
  Methexp cluster, set the right hostname or IP for \code{master} (i.e. your IP), and
  the cluster computers (these should be set to the correct static IP by default). If
  your \pkg{drake} targets perform parallel computations (e.g., MCMC chain parallelization
  in \pkg{rstan}), set \code{cores = 1L} in \code{methexp_cluster()} to avoid spawning
  too many jobs due to the nested parallelization. Make sure that all cluster computers
  have \pkg{future} and \pkg{drake} installed, otherwise you'll get hard-to-understand
  errors about missing packages or scheduler problems.

  Finally, if you use \pkg{rstan}, you may want to compile your models first and sample
  in separate targets. If you parallelize on the Methexp cluster, you may want to compile
  locally by setting \code{hpc = FALSE} in the \pkg{drake} plan and set
  \code{rstan_options(auto_write = TRUE)} to avoid unnecessary model recompilation.
}
