% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_new_project.R
\name{mk_new_project}
\alias{mk_new_project}
\title{mk_new_project}
\usage{
mk_new_project(path, ...)
}
\arguments{
\item{path}{The path to the new project directory.}
}
\value{
Invisible. Used by RStudio to set up a new project using the template.
}
\description{
Creates a folder structure for a new project according to the template.
}
