\name{[MODULE]-package}
\alias{[MODULE]-package}
\alias{[MODULE]}
\docType{package}
\title{
Simulate PK
}
\description{
Simulate PK
}
\details{
\tabular{ll}{
Package: \tab [MODULE]\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
License: \tab MIT (>= 2)\cr
}
}
\author{
InsightRX

Maintainer: InsightRX
}
\references{
}
\keyword{ package }
