% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{path}
\alias{path}
\title{Concatenate strings to file path}
\usage{
path(...)
}
\arguments{
\item{...}{one or more \R objects, to be converted to character vectors.}
}
\description{
Concatenate vectors to file path after converting to character.
}
\examples{
path("foo", "bar")
}
