% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{promise_decorate}
\alias{promise_decorate}
\title{Decorate function with future_promise}
\usage{
promise_decorate(fun)
}
\arguments{
\item{fun}{Function to decorate}
}
\value{
The input function decorated with future_promise
}
\description{
Decorate function with future_promise
}
