% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_api.R
\name{run_api}
\alias{run_api}
\title{Run plumber object}
\usage{
run_api(..., pr = prepare_api())
}
\arguments{
\item{...}{All parameters for run API.}

\item{pr}{Plumber object by default prepare_api()}
}
\description{
Run plumber object
}
\examples{
run_api()
}
