\name{template-package}
\alias{template-package}
\alias{template}
\docType{package}
\title{\packageTitle{template}}
\description{\packageDescription{template}}
\section{Package Content}{\packageIndices{template}}
\author{\packageAuthor{template}}
\section{Maintainer}{\packageMaintainer{template}}
\keyword{package}
