% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{author_from_desc}
\alias{author_from_desc}
\title{#' Initialize \pkg{target} workflow
#'
#' Sets up build files and creates a dedicated \pkg{target} cache.
#'
#' @param x Character. Location at which to initialize \pkg{target} workflow.
#' @inheritParams init_project
#'
#' @return
#' @export}
\usage{
author_from_desc(...)
}
\arguments{
\item{...}{Passed to \code{desc::desc_get}}
}
\value{

}
\description{

}
\details{

}
