% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolve2.R
\name{convolve2}
\alias{convolve2}
\title{Convolve Two Vectors}
\usage{
convolve2(a, b)
}
\arguments{
\item{a}{A numeric vector.}

\item{b}{A numeric vector.}
}
\value{
A numeric vector.
}
\description{
This function demonstrates a Rust implementation of the convolution algorithm
shown in Section 5.10.1 "Calling .Call" of \href{https://cran.r-project.org/doc/manuals/R-exts.html#Calling-_002eCall}{\emph{Writing R Extensions}}.
}
\examples{
convolve2(c(1,2,3), c(2,4,8))
}
