% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template-foo.R
\name{foo}
\alias{foo}
\title{foo}
\usage{
foo(x, y)
}
\arguments{
\item{x}{Numeric.}

\item{y}{Numeric.}
}
\description{
Sum function.
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Template}
\keyword{template}
