% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello World!}
\usage{
hello(txt = "world")
}
\arguments{
\item{txt}{Character (length = 1).
This argument will print the word input after the word, "Hello"}
}
\description{
This function will print out the words, "Hello
world"
}
\examples{
# Print the word "Hello, y'all"
hello(txt = "y'all")

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
