% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_project.R
\name{init_project}
\alias{init_project}
\title{Set up new project repository}
\usage{
init_project(x, path = ".", git = TRUE, pkg_structure = FALSE,
  packrat = TRUE)
}
\arguments{
\item{x}{Character. Name of project repository.}

\item{path}{Character. Path to the folder that is to contain the project repository.}

\item{git}{Logical. Whether to initialize git.}

\item{pkg_structure}{Logical. Whether to add R-package infrastructure.}

\item{packrat}{Logical. Whether to use \pkg{packrat} to manage package dependencies.}
}
\description{
Creates a folder structure and sets up basic functionality.
}
