% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{init_targets}
\alias{init_targets}
\title{Initialize \pkg{targets} workflow}
\usage{
init_targets(x, path = ".", git, pkg_structure)
}
\arguments{
\item{x}{Character. Location at which to initialize \pkg{targets} workflow.}

\item{path}{A path. If it exists, it is used. If it does not exist, it is
created, provided that the parent path exists.}

\item{git}{Logical. Whether to initialize git.}

\item{pkg_structure}{Logical. Whether to add R-package infrastructure.}
}
\value{

}
\description{
Sets up build files and creates a dedicated \pkg{targets} cache.
}
