% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methexp_cluster.R
\name{methexp_cluster}
\alias{methexp_cluster}
\title{Create a Parallel Socket Cluster}
\usage{
methexp_cluster(
  user = "computer",
  master = "134.95.17.36",
  servants = paste0("134.95.17.", 62:65),
  cores = 6L
)
}
\arguments{
\item{user}{Character. SSH user name on lab machines, needs to be created by admin.}

\item{master}{Character. IPv4 address of your own machine.}

\item{cores}{Integer. The number of cores per servant host.}
}
\description{
This convenience function creates a PSOCK cluster on our spare lab machines.
After finishing your computations, remember to \code{\link{parallel::stopCluster()}}.
}
\details{
Before you can use the cluster, you need to copy your SSH key to the lab machines.
Ask \code{marius.barth@uni-koeln.de} for help.
}
