% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrnorm.R
\name{myrnorm}
\alias{myrnorm}
\title{Sample from the Normal Distribution}
\usage{
myrnorm(n, mean, sd)
}
\arguments{
\item{n}{An integer for the number of random samples.}

\item{mean}{A scalar giving the mean.}

\item{sd}{A scalar giving the standard deviation.}
}
\value{
A numeric vector.
}
\description{
This is a simple little function that demonstrates calling a 'Rust' function.
It generates samples from the normal distribution.
}
\examples{
myrnorm(10, 5, 4)
}
