% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template-package.R
\docType{package}
\name{template-package}
\alias{template-package}
\title{template is a forecaster package}
\description{
The template forecaster
}
\author{
firstname lastname
Maintainer: first_name last_name email
}
\keyword{package}
