% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template-package.R
\docType{package}
\name{template-package}
\alias{template}
\alias{template-package}
\title{template: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FlandrinU/template}
  \item Report bugs at \url{https://github.com/FlandrinU/template/issues}
}

}
\author{
\strong{Maintainer}: Ulysse Flandrin \email{ulysse.flandrin@gmail.com} [copyright holder]

}
\keyword{internal}
