% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template-package.R
\docType{package}
\name{template-package}
\alias{template}
\alias{template-package}
\title{template: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Loic-sanchez/template}
  \item Report bugs at \url{https://github.com/Loic-sanchez/template/issues}
}

}
\author{
\strong{Maintainer}: Loic Sanchez \email{loic.sanchez.ls@gmail.com} (\href{https://orcid.org/0000-0002-3795-2641}{ORCID}) [copyright holder]

}
\keyword{internal}
