% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template-package.R
\docType{package}
\name{template-package}
\alias{template-package}
\title{template package}
\description{
Generic template for research projects structured as R packages.
This package works as a template for new research projects,
under the idea of having everything (data, R scripts, functions and manuscripts)
contained in the same package to promote reproducibility.
}
\details{
For detailed steps on how to use the package see \url{https://github.com/Pakillo/template/blob/master/README.md}.
}
\author{
Francisco Rodriguez-Sanchez
}
\keyword{package}

