% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/template-package.R
\docType{package}
\name{template-package}
\alias{template}
\alias{template-package}
\title{template}
\description{
Brief summary of the template package
}
\details{
\tabular{ll}{
Package: \tab template\cr
Type: \tab Package\cr
Version: \tab See sessionInfo() or DESCRIPTION file\cr
Date: \tab 2014-2015\cr
License: \tab BSD-2-clause\cr
LazyLoad: \tab yes\cr
}
}
\examples{
library(template)
}
\author{
Leo Lahti, \email{leo.lahti@iki.fi}
}
\references{
See citation("template")
\url{https://github.com/ropengov/template}
}
\keyword{package}

